Added Climbing.java
It contains a fake simulation of climbing, beware that this code isn't compileable, I just wrote it for fun when I saw your t-shirt shop. If you want to make it work, someine needs to implement methods: reachedTop and chalkHands.
Also beware that the hands never actually gets sweaty in this fake simulation.
If someone is reading this, I hope you're having a wonderful day! :)